Fix install script.
Also some cleanup.

+++ vagrant/create_scion_box.sh
@@ -3,44 +3,62 @@
 set -e
 
 VERSION="0.0.1.2"
-
 BASE=$(realpath $(dirname "$0"))
 cd "$BASE"
 
+do_package=0
+usage="$(basename "$0") [-p]
+
+where:
+    -h          this help
+    -p          also package the VM"
+
+while getopts ":hp" opt; do
+  case $opt in
+    p)
+      do_package=1
+      ;;
+    h)
+      echo "$usage" >&2
+      exit 0
+      ;;
+    \?)
+      echo "Invalid option: -$OPTARG" >&2
+      echo "$usage" >&2
+      exit 1
+      ;;
+    :)
+      echo "Option -$OPTARG requires an argument." >&2
+      echo "$usage" >&2
+      exit 1
+      ;;
+  esac
+done
 
 # running this script will create a new vagrant virtual machine with scion installed and ready to work
-# after this, you can run `vagrant package` to get a .box file that can be uploaded to vagrant cloud
+# if do_package==1, it will also run `vagrant package` to get a .box file that can be uploaded to vagrant cloud
 # for more info see https://www.vagrantup.com/docs/vagrant-cloud/boxes/create-version.html
 cp ../scion_install_script.sh .
 vagrant destroy -f
+VAGRANT_VAGRANTFILE=Vagrantfile-bootstrap vagrant destroy -f
 echo '------------------------------------ updating vagrant boxes'
 vagrant box update
-echo '------------------------------------ creating vagrant VM OS'
-VAGRANT_VAGRANTFILE=Vagrantfile-base vagrant up
-VAGRANT_VAGRANTFILE=Vagrantfile-base vagrant halt
-echo '------------------------------------ creating vagrant VM base'
+echo '------------------------------------ creating bootstrap vagrant VM'
+VAGRANT_VAGRANTFILE=Vagrantfile-bootstrap vagrant up
+echo '------------------------------------ creating base vagrant VM'
 vagrant up
-echo '------------------------------------ packaging vagrant VM'
-rm -f scion-base.box
-# vagrant package --base SCIONLabVM-base --output scion-base.box --vagrantfile Vagrantfile-exported
-vagrant package --output scion-base.box --vagrantfile Vagrantfile-exported
-vagrant destroy -f
 
-
-echo '------------------------------------ locally adding vagrant VM'
-vagrant box remove scion/ubuntu-16.04-64-scion --box-version "$VERSION" || true
-# vagrant box add "scion/ubuntu-16.04-64-scion" scion-base.box
-
-cp metadata-template.json metadata.json
-sed -i -- "s|_VERSION_|$VERSION|g" metadata.json
-sed -i -- "s|_DIRFULLPATH_|$BASE|g" metadata.json
-vagrant box add metadata.json
-echo '------------------------------------'
+if [ $do_package -eq 1 ]; then
+    echo '------------------------------------ packaging base vagrant VM'
+    rm -f scion-base.box
+    vagrant package --output scion-base.box --vagrantfile Vagrantfile-exported
+    vagrant destroy -f
+    echo '------------------------------------ locally adding base vagrant VM'
+    vagrant box remove -f scion/ubuntu-16.04-64-scion --box-version "$VERSION" || true
+    cp metadata-template.json metadata.json
+    sed -i -- "s|_VERSION_|$VERSION|g" metadata.json
+    sed -i -- "s|_DIRFULLPATH_|$BASE|g" metadata.json
+    vagrant box add metadata.json
+    echo '------------------------------------'
+fi
 echo "Done."
-
-
-# curl 'https://vagrantcloud.com/api/v1/box/scion/ubuntu-16.04-64-scion/version/0.0.1.2/provider/virtualbox/upload?access_token=zaGkqfsnqqXIpg.atlasv1.vRq2XBHhOcx5YU8fABDZy88mQxVcXEFxbuyzrGGRCRJJRLf3XUKbgeGxY7fV2fzkd2w'
-
-# curl -X PUT --upload-file scion-base.box 'https://archivist.vagrantup.com/v1/object/eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJrZXkiOiJib3hlcy80OTA3NzQwMi1mNTZkLTQ4Y2EtODVhMS03NTAyZDRiNGViZDgiLCJtb2RlIjoidyIsImV4cGlyZSI6MTUzNTMwNDY3NiwiY2FsbGJhY2siOiJodHRwczovL3ZhZ3JhbnRjbG91ZC5jb20vYXBpL2ludGVybmFsL2FyY2hpdmlzdC9jYWxsYmFjayJ9.JyhdnkITP8HcwF1wlp48cs9EX5MPIMewqX3ibCxIz-Q'
-
-# curl 'https://vagrantcloud.com/api/v1/box/scion/ubuntu-16.04-64-scion/version/0.0.1.2/provider/virtualbox?access_token=zaGkqfsnqqXIpg.atlasv1.vRq2XBHhOcx5YU8fABDZy88mQxVcXEFxbuyzrGGRCRJJRLf3XUKbgeGxY7fV2fzkd2w'
